refactor: organize test files alongside their components
- Move SidebarToggle.test.tsx to base/__tests__/ folder
- Move story-related test files to story/__tests__/ folder
- Update all import paths in moved test files
- Tests now organized alongside the components they test
- Improved maintainability and clear component-test relationships
