+++ src/components/sidebar/base/__tests__/SidebarToggle.test.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { render, screen, fireEvent } from '@testing-library/react';
 import { vi } from 'vitest';
-import SidebarToggle from '../base/SidebarToggle';
-import { setupSidebarMocks, resetSidebarMocks, mockT } from './sidebarMocks';
+import SidebarToggle from '../SidebarToggle';
+import { setupSidebarMocks, resetSidebarMocks, mockT } from '../../__tests__/sidebarMocks';
 
 // Setup mocks before tests
 setupSidebarMocks();
