+++ src/components/sidebar/story/__tests__/StoriesSection.test.tsx
@@ -9,7 +9,7 @@ import {
   mockSampleStories,
   mockUser,
   mockT,
-} from './sidebarMocks';
+} from '../../__tests__/sidebarMocks';
 import type { DatabaseSavedTranslationWithDetails } from '../../../types/database/translation';
 import type { DifficultyLevel } from '../../../types/llm/prompts';
