+++ src/components/sidebar/story/__tests__/VocabularySection.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { render, screen } from '@testing-library/react';
 import { vi } from 'vitest';
 import VocabularySection from '../VocabularySection';
-import { setupSidebarMocks, resetSidebarMocks } from './sidebarMocks';
+import { setupSidebarMocks, resetSidebarMocks } from '../../__tests__/sidebarMocks';
 
 type VocabularySidebarProps = {
   currentLanguageId?: number;
